requirements.txt
flask==0.12
flask_httauth==3.2.2
flask_restful==0.3.5
peewee==2.8.8
argon2-cffi==16.3.0

